requirements.txt
fastapi
uvicorn
fastapi-mail
python-dotenv
pydantic-settings